requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
attrs==25.3.0
audioop-lts==0.2.1
beautifulsoup4==4.13.4
certifi==2025.1.31
charset-normalizer==3.4.1
discord.py==2.5.2
frozenlist==1.6.0
idna==3.10
multidict==6.4.3
propcache==0.3.1
psycopg2-binary==2.9.10
python-dotenv==1.1.0
requests==2.32.3
soupsieve==2.7
typing_extensions==4.13.2
urllib3==2.4.0
yarl==1.20.0
